Fixes #43604 weird markup in page after remove formatting
Iterating with for (var i in array) will iterate over all properties of
the array, including all properties inherited through the prototype.  If
the prototype is extended (Array.prototyp.indexOf = function(){...})
this will also cause the stringified function source code to be an
element of the array. This caused an element to be inserted into the
page where the tag and attribute names were the functions source code,
e.g.

<function(needle){for(var i;return-1;}=\"\" return=\"\" if(this[i]=\"==needle)\" i=\"0,l=this.length;i&lt;l;i++)\"></function(needle){for(var>

If one iterates over arrays, always the following ideom is preferred:
for(var i = 0; i < x; i++)

If one wants to iterate over the properties of an object with for (var i
in x), one has to make sure to exclude the properties inherited through
the prototype (if that's what one wants).